+++ kpi-tracker/CHANGELOG.md
@@ -8,6 +8,8 @@
 - Add graph showing active staking accounts, divided into bakers and delegators.
 - Add graph showing status of active bakers.
 - Add graph of delegation recipients.
+- Add graph for minted CCDs.
+- Add graphs for baker, finalizer, and foundation rewards.
 
 ## 1.1.0